+++ engine/lib/phases/phase_sort_items_namespace_wise.ml
@@ -16,5 +16,5 @@ module Make (F : Features.T) =
 
       let ditems items =
         let module Deps = Dependencies.Make (F) in
-        Deps.sort items
+        Deps.sort_namespace_wise items
     end)