+++ hax-lib/core-models/src/core/array.rs
@@ -49,6 +49,7 @@ pub fn from_fn<T, const N: usize, F: crate::ops::function::FnOnce<usize, Output 
     array_from_fn(f)
 }
 
+#[cfg_attr(hax_backend_lean, hax_lib::exclude)]
 impl<T, const N: usize> crate::iter::traits::collect::IntoIterator for [T; N] {
     type IntoIter = iter::IntoIter<T, N>;
     fn into_iter(self) -> iter::IntoIter<T, N> {
@@ -62,6 +63,7 @@ use crate::ops::{
 };
 
 #[hax_lib::attributes]
+#[cfg_attr(hax_backend_lean, hax_lib::exclude)]
 impl<T, const N: usize> Index<usize> for [T; N] {
     type Output = T;
     #[hax_lib::requires(i < self.len())]
@@ -71,6 +73,7 @@ impl<T, const N: usize> Index<usize> for [T; N] {
 }
 
 #[hax_lib::attributes]
+#[cfg_attr(hax_backend_lean, hax_lib::exclude)]
 impl<T, const N: usize> Index<Range<usize>> for [T; N] {
     type Output = [T];
     #[hax_lib::requires(i.start <= i.end && i.end <= self.len())]
@@ -79,6 +82,7 @@ impl<T, const N: usize> Index<Range<usize>> for [T; N] {
     }
 }
 #[hax_lib::attributes]
+#[cfg_attr(hax_backend_lean, hax_lib::exclude)]
 impl<T, const N: usize> Index<RangeTo<usize>> for [T; N] {
     type Output = [T];
     #[hax_lib::requires(i.end <= self.len())]
@@ -87,6 +91,7 @@ impl<T, const N: usize> Index<RangeTo<usize>> for [T; N] {
     }
 }
 #[hax_lib::attributes]
+#[cfg_attr(hax_backend_lean, hax_lib::exclude)]
 impl<T, const N: usize> Index<RangeFrom<usize>> for [T; N] {
     type Output = [T];
     #[hax_lib::requires(i.start <= self.len())]
@@ -95,6 +100,7 @@ impl<T, const N: usize> Index<RangeFrom<usize>> for [T; N] {
     }
 }
 #[hax_lib::attributes]
+#[cfg_attr(hax_backend_lean, hax_lib::exclude)]
 impl<T, const N: usize> Index<RangeFull> for [T; N] {
     type Output = [T];
     fn index(&self, i: RangeFull) -> &[T] {
@@ -106,6 +112,7 @@ mod iter {
     use crate::option::Option;
     use rust_primitives::sequence::*;
     pub struct IntoIter<T, const N: usize>(pub Seq<T>);
+    #[cfg_attr(hax_backend_lean, hax_lib::exclude)]
     impl<T, const N: usize> crate::iter::traits::iterator::Iterator for IntoIter<T, N> {
         type Item = T;
         fn next(&mut self) -> Option<T> {