+++ CHANGELOG.md
@@ -65,7 +65,7 @@ Changes to the Lean backend:
  - Add generation of specs from requires/ensures-annotations (#1815)
  - Add support for nonliteral array sizes (#1826)
  - Add `hax_lib::lean::proof` attribute (#1831)
- - Add support for `#[hax_lib::opaque]` (#1846)
+ - Add support for `#[hax_lib::opaque]` (#1846, #1887)
  - Turn rejection phase into a transformation phase (#1840)
 
 Miscellaneous: